chore: moved upgrades into stage1Upgrades
