+++ src/genericRobot.cpp
@@ -7,7 +7,7 @@
 #include "genericRobot.h"
 #include "abstractRobot/robot.h"
 #include "logger.h"
-#include "upgrades/upgrades.h"
+#include "stage1Upgrades/upgrades.h"
 #include "vector2d.h"
 #include "environment.h"
