+++ src/genericRobot.h
@@ -10,7 +10,7 @@
 
 #include "vector2d.h"
 #include "logger.h"
-#include "upgrades/upgrades.h"
+#include "stage1Upgrades/upgrades.h"
 
 #include <vector>
 #include <cstdint>
