+++ src/stage1Upgrades/hideBot.cpp
