+++ src/stage1Upgrades/hideBot.h
