+++ src/stage1Upgrades/scoutBot.cpp
