+++ src/stage1Upgrades/thirtyShotBot.h
